clean code: Move Netex classes into loader, index and mapping packages based on where they belong.
