+++ src/main/java/org/opentripplanner/netex/mapping/NetexMapper.java
@@ -15,8 +15,8 @@ import org.opentripplanner.model.Trip;
 import org.opentripplanner.model.TripPattern;
 import org.opentripplanner.model.calendar.ServiceCalendarDate;
 import org.opentripplanner.model.impl.OtpTransitServiceBuilder;
-import org.opentripplanner.netex.loader.NetexEntityDataIndex;
-import org.opentripplanner.netex.loader.NetexImportDataIndexReadOnlyView;
+import org.opentripplanner.netex.index.NetexEntityDataIndex;
+import org.opentripplanner.netex.index.NetexImportDataIndexReadOnlyView;
 import org.opentripplanner.netex.loader.mapping.FlexStopLocationMapper;
 import org.opentripplanner.netex.mapping.calendar.CalendarMapper;
 import org.opentripplanner.netex.support.DayTypeRefsToServiceIdAdapter;
